libvirt: fix bug due to a static variable
connection was shared among all the instances of the libvirt manager, it was
done withour the required synchornization mechanishms, this leads to
random failures.

There is not any particular reason to have that variabled shared among all the
instances, then this commit modifies it in a way that the variable is local to
each instance.